Criando funcionalidades de dashboard e relatórios do projeto

+++ back-end/src/main/java/com/f1/api/domain/escuderia/Escuderia.java
@@ -1,5 +1,6 @@
 package com.f1.api.domain.escuderia;
 
+import com.f1.api.dto.escuderia.DadosCadastroEscuderia;
 import jakarta.persistence.*;
 
 import java.util.Objects;
@@ -11,7 +12,7 @@ public class Escuderia {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "constructorid")
-    private Long id;
+    private Integer id;
 
     @Column(name = "constructorref", nullable = false)
     private String referencia;
@@ -35,11 +36,11 @@ public class Escuderia {
         this.url = dados.url();
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
